+++ hyperframe/frame.py
@@ -485,6 +485,12 @@ class PushPromiseFrame(Padding, Frame):
         )
         self.body_len = len(data)
 
+        if self.promised_stream_id == 0 or self.promised_stream_id % 2 != 0:
+            raise InvalidFrameError(
+                "Invalid PUSH_PROMISE promised stream id: %s" %
+                self.promised_stream_id
+            )
+
         if self.pad_length and self.pad_length >= self.body_len:
             raise InvalidPaddingError("Padding is too long.")